+++ packages/vite-plugin-vui/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vite-plugin-vui
 
+## 1.1.92
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.18.1
+
 ## 1.1.91
 
 ### Patch Changes