+++ packages/vui-wysiwyg/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @fastkit/vui-wysiwyg
 
+## 6.0.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fastkit/vui@0.18.1
+
 ## 6.0.0
 
 ### Minor Changes